+++ src/event/MainEventLoop.h
@@ -18,8 +18,8 @@ public:
     virtual void post(std::unique_ptr<Event>&& event) override;
 
     using EventLoop::startTimer;
-    virtual uint64_t startTimer(const std::shared_ptr<Event>& event, uint64_t timeout, TimerMode mode = TimerMode::SingleShot) override;
-    virtual void stopTimer(uint64_t id) override;
+    virtual uint32_t startTimer(const std::shared_ptr<Event>& event, uint64_t timeout, TimerMode mode = TimerMode::SingleShot) override;
+    virtual void stopTimer(uint32_t id) override;
 
     virtual void* handle() const override;
